Decouple usage measurement from billing
- Create new `engine/core-modules/usage/` module with its own resolver,
  services, DTOs, enums, types, and listener
- Move all usage-related code out of billing module:
  constants, enums, types, DTOs, analytics service, event writer, listener
- Split listener: UsageEventListener (ClickHouse writes) stays in usage,
  BillingUsageEventListener (Stripe metering) stays in billing
- Extract getUsageAnalytics query into UsageResolver with WORKSPACE
  permission (not BILLING) since usage is available to self-hosters
- Remove billUsage delegation from UsageEventWriterService — billing
  module handles its own Stripe metering directly
- Clean up billing.resolver.ts: remove unused imports, repository
  injection, and helper methods
- Change quantity from Float64 to Int64 in ClickHouse
- Fix mapToDisplayCredits naming: SQL now aliases as creditsUsedMicro
  with explicit typed intermediate (BreakdownRowMicro/TimeSeriesRowMicro)
  before converting to display credits

Made-with: Cursor

+++ packages/twenty-server/src/database/clickHouse/migrations/004-create-usage-event-table.sql
@@ -5,7 +5,7 @@ CREATE TABLE IF NOT EXISTS usageEvent
     `userWorkspaceId`    String DEFAULT '',
     `resourceType`       LowCardinality(String) NOT NULL,
     `operationType`      LowCardinality(String) NOT NULL,
-    `quantity`           Float64 NOT NULL DEFAULT 0,
+    `quantity`           Int64 NOT NULL DEFAULT 0,
     `unit`               LowCardinality(String) NOT NULL DEFAULT 'CREDIT',
     `creditsUsedMicro`   Int64 NOT NULL DEFAULT 0,
     `resourceId`         String DEFAULT '',
